requirements.txt
asyncpg
humanize
discord.py
aiohttp